The length of array like objects can in some cases be 2^53 -1 in ES7
Add a Value::getLength(), that converts a Value to a length bound
between 0 and 2^53-1 as per ES7 spec. Use the extended range in
Array.prototype.splice and map to fix hanging test cases.

+++ src/qml/jsruntime/qv4argumentsobject_p.h
@@ -149,7 +149,7 @@ struct ArgumentsObject: Object {
     static bool putIndexed(Managed *m, uint index, const Value &value);
     static bool deleteIndexedProperty(Managed *m, uint index);
     static PropertyAttributes queryIndexed(const Managed *m, uint index);
-    static uint getLength(const Managed *m);
+    static qint64 getLength(const Managed *m);
 
     void fullyCreate();
